+++ astronz/agn.py
@@ -3,108 +3,23 @@
 # ASTRONZ: astro calc utilities
 
 
-import sys,string,os,math,numpy
-
-RAD2DEG=180./math.pi
-HI=1.42040575177e+09 #GHz
-TSPIN=100 #K
-MSUN=1.98855e33
-MHI=1.6749E-24
-CHI=2.36E5
-PC=3.08567758E18 #cm
-
-JANSKY=1e-23    #erg/scm2Hz
-
-C=2.99792458E10    #cm/s
-G=6.6742E-08       #cm3kg-1s-1      
-MP=1.67492728E-24  #g
-SIGMAT=6.66524E-25 #cm2
+import sys,string,os
+import numpy as np
 
 class AGN:
       
-  def __init__(self,args):
+  def __init__(self):
 
-# elif (inp=='A' or inp=='a' or inp=='AGN'):
-#   in1="\nHere's what I can do:\n"
-#   in2="*Eddington(e)\t*Mechanical Luminosity (LM)\t*Radiative Luminosity(LR)\n*Lambda(LA)\t*Bondi(B)\t*X-accretion\n"
-  
-#   inp=str(raw_input(in1+in2))
-  
-#   if (inp=='e' or inp=='E' or inp=='eddington'):
-#       in1="\nM_BH(M_sun)= "
-#       m=a.eddington(float(raw_input(in1)))
-#       print "Medd = %g M_sun/yr\n"% (m[0])
-#       print "Ledd = %g erg/s"% (m[1])
-#       print "Ledd1 = %g M_sun/yr"% (m[2])
-#   elif (inp=='lm' or inp=='Mechanical Luminosity' or inp=='LM'):
-#       in1="\nz = "
-#       in2="F (Jy)= "
-#       in3="M_BH (M_sun)= "
-#       z=float(raw_input(in1))
-#       f=float(raw_input(in2))
-#       mbh=float(raw_input(in3))
-#       lmech=a.mechanical(z,f,mbh)
-#       print "L_mech = %g erg/s\n"% (lmech[0])
-#       print "L_mech/Ledd= %g "%(lmech[1])
-#   elif (inp=='lr' or inp=='Radiative Luminosity' or inp=='LR'):
-#       in1="\nz = "
-#       in2="F_OIII (erg/scm2)= "
-#       in3="M_BH (M_sun)= "
-#       z=float(raw_input(in1))
-#       f=float(raw_input(in2))
-#       mbh=float(raw_input(in3))
-#       lmech=a.radiative(z,f,mbh)
-#       print "L_rad = %g erg/s\n"% (lmech[0])
-#       print "L_rad/Ledd= %g "%(lmech[1])
-#   elif (inp=='LA' or inp=='lambda' or inp=='la'):
-#       in1="\nz = "
-#       in2="F_OIII (erg/scm2)= "
-#       in3="F_1.4 (Jy)= "
-#       in4="M_BH (M_sun)= "
-#       z=float(raw_input(in1))
-#       foiii=float(raw_input(in2))
-#       f=float(raw_input(in3))
-#       mbh=float(raw_input(in4))
-#       lmech=a.lamb(z,f,foiii,mbh)
-#       print "Lambda = %g \n"% (lmech[0])
-#       print "LOGLambda = %g \n"% (lmech[1])
-#   elif (inp=='b' or inp=='Bondi' or inp=='B'):
-#       in1="\nz = "
-#       in2="F (Jy)= "
-#       in3="M_BH (M_sun)= "
-#       z=float(raw_input(in1))
-#       f=float(raw_input(in2))
-#       mbh=float(raw_input(in3))
-#       pj=a.bondi(z,f,mbh)
-#       print "P_jet = %g erg/s\n"% (pj[0])
-#       print "P_Bondi = %g erg/s\n"% (pj[1])
-#       print "M_Bondi(B) = %g M_sun/yr\n"% (pj[2])
-#       print "M_Bondi_jet(B) = %g M_sun/yr\n"% (pj[3])
-#       print "M_Bondi(A) = %g M_sun/yr\n"% (pj[4])
-#       print "M_bondi(BH) = %g\n"% (pj[5])
-#       print "eff= %g\n"% (pj[6])
-#       print "cloud= %g\n"% (pj[7])
-#       print "pjet= %g\n"% (pj[8])
-#       print "pallent= %g\n"% (pj[9])
-#       print "pbonditop= %g\n"% (pj[10])
-#   elif (inp=='x' or inp=='Xaccretion' or inp=='X-ray'):
-#       in1="\nz = "
-#       in2="F (Jy)= "
-#       in3="M_BH (M_sun)= "
-#       z=float(raw_input(in1))
-#       f=float(raw_input(in2))
-#       mbh=float(raw_input(in3))
-#       pj=a.Xaccretion(z,f,mbh)
-#       print "Xratio = %g \n"% (pj)
-# else:
-#   print 'you have not entered correct function\n EXIT FAILURE MOROAN'
+    print ('\n\t************* --- ASTRONZ : AGN --- **************\n')
+    print ('\t\t   ... AGN Science tools ... \n')
 
-    self.c=C
-    self.G=G
-    self.mp=MP
-    self.sigmaT=SIGMAT
-    self.msun=MSUN
-    self.myr=(3600*365*24)/MSUN
+    self.rad2deg = 180./np.pi
+    self.c=2.99792458E10    #cm/s
+    self.G=6.6742E-08       #cm3kg-1s-1
+    self.mp=1.67492728E-24  #g
+    self.sigmaT=6.66524E-25 #cm2
+    self.msun=1.98855e33
+    self.myr=(3600*365*24)/self.msun
     self.yr=(3600*365*24)
     self.ledd=1.3e38
     self.rad=7.3e36
@@ -114,13 +29,13 @@ class AGN:
     self.p0=1.0e44   #erg/s
     self.l0=7e29     #erg/sHz
     self.ev= 1.602177e-12 #from ev to erg
-    self.pc=PC
+    self.pc=3.08567758E18 #cm
     
   def eddington(self,mbh):
     
     #set mbh in g:
     mbhg=mbh*self.msun
-    a=4*math.pi*self.G*self.mp*self.c
+    a=4*np.pi*self.G*self.mp*self.c
     ledd=a*mbhg/self.sigmaT
     medd=ledd/(self.eff*(self.c)**2)*self.myr
 
@@ -133,21 +48,21 @@ class AGN:
     
     lum=c.luminosity(z,f)*1e-7
     lmec=self.rad*pow(lum/1e24,0.71) *1e7       #relation in W result in erg/s
-    #a=a.eddington(mbh)[1]
+    #a=self.eddington(mbh)[1]
     
-    lmecedd=lmec/a.eddington(mbh)[1]        #in erg/s
+    lmecedd=lmec/self.eddington(mbh)[1]        #in erg/s
     
     return lmec,lmecedd
   
   def radiative(self, z,foiii,mbh):
     
     dl=c.lum_dist(z)
-    lum=(foiii*4*math.pi*(dl)**2)
+    lum=(foiii*4*np.pi*(dl)**2)
 
 
     lrad=self.oiii*lum
     
-    lradedd=lrad/a.eddington(mbh)[1]        #in erg/s
+    lradedd=lrad/self.eddington(mbh)[1]        #in erg/s
     
     return lrad, lradedd
     
@@ -155,14 +70,14 @@ class AGN:
     
     lr=self.radiative(z,foiii,mbh)[1]
     lm=self.mechanical(z,f,mbh)[1]
-    ledd=a.eddington(mbh)[1]
+    ledd=self.eddington(mbh)[1]
 
   
     lamb=(lm+lr)
   
   
   
-    loglam=math.log10(lamb)
+    loglam=np.log10(lamb)
     
     return lamb, loglam
     
@@ -184,7 +99,7 @@ class AGN:
     # from Balmaverde et al. 2008 Pjet => Pbondi = Mbondi*c^2
                                 # Pjet= 0.012 Mbondi * c^2
     
-    pbondi=(math.log10(pjet/1e43)+1.91)/1.1   #the formula is in 1e43 units
+    pbondi=(np.log10(pjet/1e43)+1.91)/1.1   #the formula is in 1e43 units
     pbondi=pow(10,pbondi)
     pbondi=pbondi*1e43
 
@@ -193,7 +108,7 @@ class AGN:
     mbondijet=pjet/((self.c**2)*0.012)*self.myr
    
     # from Allen et al. 2006 Pjet => Pbondi = 0.1 * Mbondi *c^2
-    pallen=(math.log10(pjet/1e43)*0.77)+0.65
+    pallen=(np.log10(pjet/1e43)*0.77)+0.65
     pallen=pow(10,pallen)
     pallen=pallen*1e43
     
@@ -204,7 +119,7 @@ class AGN:
   
     #determine c_s
     kbt=0.7e3*self.ev
-    cs=math.sqrt((5/3*kbt)/(0.62*self.mp))
+    cs=np.sqrt((5/3*kbt)/(0.62*self.mp))
     
     #determine the density of the gas
     ne=1
@@ -220,7 +135,7 @@ class AGN:
     lambdaa=0.25
     eff=0.1
 
-    bondi=lambdaa*math.pi*cs*ra**2*rho*self.myr
+    bondi=lambdaa*np.pi*cs*ra**2*rho*self.myr
     
     pbonditop=0.017*bondi/self.myr*self.c**2
 
@@ -230,7 +145,7 @@ class AGN:
     #general relativity
     #you confront the mass accretion rate with the eddington luminosity
     
-    ledd=a.eddington(mbh)[1]
+    ledd=self.eddington(mbh)[1]
     ledduni=ledd/self.c**2*self.myr
     
     accr=bondi/ledduni
@@ -254,7 +169,7 @@ class AGN:
     press=kbt*T*rho/(0.62*self.mp)
     print press
     
-    V=math.pi*minorax*majorax**2
+    V=np.pi*minorax*majorax**2
     print V
 
     tage=1e2*self.yr
@@ -266,23 +181,106 @@ class AGN:
     accrcloud=mcloud*self.msun/t*self.myr
     
     # from Allen et al. 2006 Pjet => Pbondi = 0.1 * Mbondi *c^2
-    pallen2=(math.log10(pjetta/1e43)*0.77)+0.65
-    pallen2=pow(10,pallen2)
+    pallen2=(np.log10(pjetta/1e43)*0.77)+0.65
+    pallen2=np.power(10,pallen2)
     pallen2=pallen2*1e43
     
     mbondia2=pallen2/((self.c**2)*1)*self.myr
-
-    
-    
 
     return pjet,pbondi,mbondi,mbondijet,mbondia,bondi,accr,accrcloud,pjetta,mbondia2,pbonditop
     
     
   def Xaccretion (self,z,f,mbh):
     
-    ledd=a.eddington(mbh)[1]
+    ledd=self.eddington(mbh)[1]
     lx=c.luminosity(z,f)
     lx=1e41
-    lxratio=math.log10(lx/ledd)
+    lxratio=np.log10(lx/ledd)
 
     return lxratio
+
+  def main(self):
+
+    in1="\t... Here's what I can do: ...\n"
+    in2='''\n\t - Eddington\t->\tE
+\t - Mech. Lum.\t->\tLM
+\t - Rad. Lum.\t->\tLR
+\t - Lambda\t->\tLA
+\t - Bondi\t->\tB
+\t - X-accretion\t->\tX\n
+      '''
+
+    inp=str(raw_input(in1+in2))
+  
+    if (inp=='e' or inp=='E' or inp=='eddington'):
+        in1="\nM_BH(M_sun)= "
+        m=self.eddington(float(raw_input(in1)))
+        print "Medd = %g M_sun/yr\n"% (m[0])
+        print "Ledd = %g erg/s"% (m[1])
+        print "Ledd1 = %g M_sun/yr"% (m[2])
+    elif (inp=='lm' or inp=='Mechanical Luminosity' or inp=='LM'):
+        in1="\nz = "
+        in2="F (Jy)= "
+        in3="M_BH (M_sun)= "
+        z=float(raw_input(in1))
+        f=float(raw_input(in2))
+        mbh=float(raw_input(in3))
+        lmech=self.mechanical(z,f,mbh)
+        print "L_mech = %g erg/s\n"% (lmech[0])
+        print "L_mech/Ledd= %g "%(lmech[1])
+    elif (inp=='lr' or inp=='Radiative Luminosity' or inp=='LR'):
+        in1="\nz = "
+        in2="F_OIII (erg/scm2)= "
+        in3="M_BH (M_sun)= "
+        z=float(raw_input(in1))
+        f=float(raw_input(in2))
+        mbh=float(raw_input(in3))
+        lmech=self.radiative(z,f,mbh)
+        print "L_rad = %g erg/s\n"% (lmech[0])
+        print "L_rad/Ledd= %g "%(lmech[1])
+    elif (inp=='LA' or inp=='lambda' or inp=='la'):
+        in1="\nz = "
+        in2="F_OIII (erg/scm2)= "
+        in3="F_1.4 (Jy)= "
+        in4="M_BH (M_sun)= "
+        z=float(raw_input(in1))
+        foiii=float(raw_input(in2))
+        f=float(raw_input(in3))
+        mbh=float(raw_input(in4))
+        lmech=self.lamb(z,f,foiii,mbh)
+        print "Lambda = %g \n"% (lmech[0])
+        print "LOGLambda = %g \n"% (lmech[1])
+    elif (inp=='b' or inp=='Bondi' or inp=='B'):
+        in1="\nz = "
+        in2="F (Jy)= "
+        in3="M_BH (M_sun)= "
+        z=float(raw_input(in1))
+        f=float(raw_input(in2))
+        mbh=float(raw_input(in3))
+        pj=self.bondi(z,f,mbh)
+        print "P_jet = %g erg/s\n"% (pj[0])
+        print "P_Bondi = %g erg/s\n"% (pj[1])
+        print "M_Bondi(B) = %g M_sun/yr\n"% (pj[2])
+        print "M_Bondi_jet(B) = %g M_sun/yr\n"% (pj[3])
+        print "M_Bondi(A) = %g M_sun/yr\n"% (pj[4])
+        print "M_bondi(BH) = %g\n"% (pj[5])
+        print "eff= %g\n"% (pj[6])
+        print "cloud= %g\n"% (pj[7])
+        print "pjet= %g\n"% (pj[8])
+        print "pallent= %g\n"% (pj[9])
+        print "pbonditop= %g\n"% (pj[10])
+    elif (inp=='x' or inp=='Xaccretion' or inp=='X-ray'):
+        in1="\nz = "
+        in2="F (Jy)= "
+        in3="M_BH (M_sun)= "
+        z=float(raw_input(in1))
+        f=float(raw_input(in2))
+        mbh=float(raw_input(in3))
+        pj=self.Xaccretion(z,f,mbh)
+        print "Xratio = %g \n"% (pj)
+    else:
+        print ('\n\t ... you have not entered correct function ... \n')
+        print ('\t************* --- AGN : ERROR --- **************\n')
+        sys.exit(0)
+
+    print ('\n\t************* --- AGN : DONE --- **************\n')
